hd_emails/requirements.txt
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
DateTime==5.5
et_xmlfile==2.0.0
idna==3.10
msal==1.32.3
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==80.3.1
six==1.17.0
smartsheet-python-sdk==3.0.5
tzdata==2025.2
urllib3==2.4.0
zope.interface==7.2
